requirements.txt
psycopg2==2.9.3
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bidict==0.22.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
comtypes==1.1.11
Deprecated==1.2.13
dnspython==2.2.1
docopt==0.6.2
eventlet==0.33.1
Flask==2.1.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.6.0
Flask-RESTful==0.3.9
Flask-Script==2.0.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
multidict==6.0.2
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pep8==1.7.1
pipwin==0.5.2
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-engineio==4.3.3
python-socketio==5.7.1
redis==4.3.4
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
